Fix invalid Target Temperature
There are at least two reports that show the adding current temperature to target temperature is not needed nor wanted. 
I think that it is because of invalid interpretation of adjustment that you can set when heatpump is using heat curve. Then targetTemp is between -5 and +5 but it doesn't mean it is currentTemp+adjustment. Actually it is f(outsideTemp)+adjustment where f(t) is the heating curve. I think even if I use heat curve (which I do in the winter) I'd like to see the adjustment I set (from -5 to +5) instead of virtual number based on current water temperature.